+++ src/parser/inline/link.ts
@@ -11,10 +11,10 @@ type SubParsers = [ImageParser] | [TextParser];
 const syntax = /^\[[^\n]*?\]\(/;
 
 export const link: LinkParser = function (source: string): Result<HTMLAnchorElement, SubParsers> {
-  if (!source.startsWith('[') || source.startsWith('[[') || !source.match(syntax)) return;
+  if (!source.startsWith('[') || !source.match(syntax)) return;
   const [first, next] = source.startsWith('[]')
     ? [[], source.slice(1)]
-    : combine<SubParsers, HTMLElement | Text>([image])(source.slice(1)) || loop(combine<SubParsers, HTMLElement | Text>([text]), /^\]|^\n/)(source.slice(1)) || [[], ''];
+    : combine<SubParsers, HTMLElement | Text>([image])(source.slice(1)) || loop(combine<SubParsers, HTMLElement | Text>([text]), /^\]\(|^\n/)(source.slice(1)) || [[], ''];
   if (!next.startsWith('](')) return;
   const children = squash(first);
   const [[, , ...second], rest] = loop(text, /^\)|^\s(?!nofollow\))/)(next) || [[], ''];