+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add insertion syntax.
 - Change strike to deletion syntax.
 - Remove `sup` and `sub` html tags.
+- Improve parser algorithm.
 
 ## 0.9.0